pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "bot"
description = "Bot aldemarromel :D"
readme = "README.md"
version = "0.1.0"
dependencies = [
	"RPi.GPIO >= 0.7.1",
	"protobuf >= 3.20.0",
	"pynmea2",
	"pyserial",
	"fastapi[standard]",
	"geojson"
]
requires-python = ">= 3.11"

[project.scripts]
